+++ test/teams.e2e-spec.ts
@@ -7,6 +7,9 @@ import { seed } from "@Prisma/seed/seed";
 import { loginAndGetTokens } from "./utils";
 import * as cookieParser from "cookie-parser";
 import { CASLForbiddenExceptionFilter } from "@/exception-filters/casl-forbidden-exception.filter";
+import { toBeArray } from "jest-extended";
+
+expect.extend({ toBeArray });
 
 //Logged in user is Jessica Williamson for admin routes /teams and /teams/voyages/:voyageid
 //Logged in user is Dan ko for team member routes /teams/:teamid and /teams/:teamid/members
@@ -49,7 +52,7 @@ describe("Teams Controller (e2e)", () => {
                 .expect(200)
                 .expect("Content-Type", /json/)
                 .expect((res) => {
-                    expect(res.body).toBeArray;
+                    expect(res.body).toBeArray();
                 });
         });
         it("should return 401 when user is not logged in", async () => {
@@ -86,7 +89,7 @@ describe("Teams Controller (e2e)", () => {
                 .expect(200)
                 .expect("Content-Type", /json/)
                 .expect((res) => {
-                    expect(res.body).toBeArray;
+                    expect(res.body).toBeArray();
                 });
         });
         it("should return 404 if voyage teams are not found given a voyage id", async () => {