+++ src/config/app/appConfig.service.ts
@@ -1,11 +1,14 @@
 import { Injectable } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
+import { readFileSync } from "fs";
+import { join } from "path";
 
 interface AppConfig {
     app: {
         nodeEnv: string;
         port: number;
         frontendUrl: string;
+        latestReleaseVersion: string;
     };
 }
 
@@ -23,5 +26,12 @@ export class AppConfigService {
         return this.configService.get<string>("app.frontendUrl", {
             infer: true,
         })!;
+    }
+
+    get latestReleaseVersion() {
+        //retrieve the latest version from package.json
+        const packageJsonPath = join(__dirname, "../../../../", "package.json");
+        const packageJson = JSON.parse(readFileSync(packageJsonPath, "utf-8"));
+        return packageJson.version;
     }
 }