+++ test/auth.e2e-spec.ts
@@ -16,6 +16,9 @@ import { CASLForbiddenExceptionFilter } from "@/exception-filters/casl-forbidden
 
 import { AuthConfig } from "@/config/auth/auth.interface";
 import { OAuthConfig } from "@/config/Oauth/oauthConfig.interface";
+import { toBeTrue, toBeFalse } from "jest-extended";
+
+expect.extend({ toBeTrue, toBeFalse });
 
 const signupUrl = "/auth/signup";
 const loginUrl = "/auth/login";
@@ -277,7 +280,7 @@ describe("AuthController e2e Tests", () => {
                 },
             });
 
-            expect(userAfterVerify?.emailVerified).toBe(true);
+            expect(userAfterVerify?.emailVerified).toBeTrue();
             expect(userAfterVerify?.emailVerificationToken).toBe(null);
         });
 
@@ -313,7 +316,7 @@ describe("AuthController e2e Tests", () => {
                 },
             });
 
-            expect(userAfterVerify?.emailVerified).toBe(false);
+            expect(userAfterVerify?.emailVerified).toBeFalse();
             expect(userAfterVerify?.emailVerificationToken).toBeDefined();
         });