+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ## [Unreleased]
 - Updated readme for installation part ([#225])(https://github.com/chingu-x/chingu-dashboard-be/pull/225)
-
+- Updated nestjs packages to latest version ([#233])(https://github.com/chingu-x/chingu-dashboard-be/pull/233)
 ### Added
 
 ### Changed