+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Changed
 
 ### Fixed
+- fixed POST voyages/teams/{teamId}/techs bug , verify that categoryId is owned by correct team ([#229](https://github.com/chingu-x/chingu-dashboard-be/pull/229))
 
 ### Removed