+++ CHANGELOG.md
@@ -24,8 +24,9 @@
 ### Changed
 
 ### Fixed
-- fixed POST voyages/teams/{teamId}/techs bug , verify that categoryId is owned by correct team ([#229](https://github.com/chingu-x/chingu-dashboard-be/pull/229))
 
+
+- fixed POST voyages/teams/{teamId}/techs bug , verify that categoryId is owned by correct team ([#229](https://github.com/chingu-x/chingu-dashboard-be/pull/229))
 ### Removed
 
 ## [v1.1.0-alpha]